Moved opening times stuff from server into common
