+++ server/services/interval-cache.js
@@ -1,6 +1,8 @@
 import IntervalCache from 'interval-cache'; // https://github.com/danneu/interval-cache
 import {getFlags} from '../services/flags-lookup';
-import {getGlobalAlert, getCollectionOpeningTimes} from '../services/prismic';
+import {getGlobalAlert} from '../services/prismic';
+// $FlowFixMe
+import {getCollectionOpeningTimes} from '../../common/services/prismic/opening-times';
 import Raven from 'raven';
 
 const fiveMinutes = 1000 * 60 * 5;
